+++ ddth-dao-core/src/test/java/com/github/ddth/dao/qnd/QndAnnotatedGenericRowMapper.java
@@ -3,7 +3,7 @@ package com.github.ddth.dao.qnd;
 import org.apache.commons.lang.builder.ToStringBuilder;
 
 import com.github.ddth.dao.BaseBo;
-import com.github.ddth.dao.jdbc.AbstractGenericBoJdbcDao;
+import com.github.ddth.dao.jdbc.GenericBoJdbcDao;
 import com.github.ddth.dao.jdbc.annotations.AnnotatedGenericRowMapper;
 import com.github.ddth.dao.jdbc.annotations.ChecksumColumn;
 import com.github.ddth.dao.jdbc.annotations.ColumnAttribute;
@@ -15,8 +15,8 @@ public class QndAnnotatedGenericRowMapper {
     public static class MyBo extends BaseBo {
     }
 
-    @PrimaryKeyColumns(columns = { "pk1", "pk2" })
-    @UpdateColumns(columns = { "col1", "col2", "col3" })
+    @PrimaryKeyColumns({ "pk1", "pk2" })
+    @UpdateColumns({ "col1", "col2", "col3" })
     @ChecksumColumn("checksum")
     @ColumnAttribute(column = "pk1", attr = "pk1", attrClass = int.class)
     @ColumnAttribute(column = "pk2", attr = "pk2", attrClass = String.class)
@@ -27,7 +27,7 @@ public class QndAnnotatedGenericRowMapper {
     public static class MyRowMapper extends AnnotatedGenericRowMapper<MyBo> {
     }
 
-    public static class MyJdbcDao extends AbstractGenericBoJdbcDao<MyBo> {
+    public static class MyJdbcDao extends GenericBoJdbcDao<MyBo> {
     }
 
     public static void main(String[] args) {
@@ -37,11 +37,15 @@ public class QndAnnotatedGenericRowMapper {
             jdbcDao.setRowMapper(rowMapper);
             jdbcDao.init();
 
-            System.out.println(new ToStringBuilder(null).append(rowMapper.getPrimaryKeyColumns()));
-            System.out.println(new ToStringBuilder(null).append(rowMapper.getInsertColumns()));
-            System.out.println(new ToStringBuilder(null).append(rowMapper.getUpdateColumns()));
-            System.out.println(new ToStringBuilder(null).append(rowMapper.getAllColumns()));
-            System.out.println(rowMapper.getChecksumColumn());
+            System.out.println("PK        : "
+                    + new ToStringBuilder(null).append(rowMapper.getPrimaryKeyColumns()));
+            System.out.println("InsertCols: "
+                    + new ToStringBuilder(null).append(rowMapper.getInsertColumns()));
+            System.out.println("UpdateCols: "
+                    + new ToStringBuilder(null).append(rowMapper.getUpdateColumns()));
+            System.out.println(
+                    "AllCols   : " + new ToStringBuilder(null).append(rowMapper.getAllColumns()));
+            System.out.println("Checksum  : " + rowMapper.getChecksumColumn());
         }
     }
 }