+++ ddth-dao-core/src/test/java/com/github/ddth/dao/test/bo/jdbc/UserBoJdbcDao.java
@@ -1,8 +1,8 @@
 package com.github.ddth.dao.test.bo.jdbc;
 
-import com.github.ddth.dao.jdbc.AbstractGenericBoJdbcDao;
+import com.github.ddth.dao.jdbc.GenericBoJdbcDao;
 
-public class UserBoJdbcDao extends AbstractGenericBoJdbcDao<UserBo> {
+public class UserBoJdbcDao extends GenericBoJdbcDao<UserBo> {
     public static void main(String[] args) {
         UserBoJdbcDao dao = new UserBoJdbcDao();
         System.out.println(dao);